Change fs.h to ustar.h for ustar related content.
